+++ public/colorsets/eco.scss
@@ -1,3 +1,6 @@
 $bg: black;
-$fc: darkgrey;
-$bc: white;
+$fc: skyblue;
+$bc: #2c3539;
+$close: darkred;
+$border: grey;
+$ld: MidnightBlue;
