+++ public/colorsets/init.scss
@@ -1,3 +1,6 @@
-$bg: black;
-$fc: darkgrey;
-$bc: white;
+$bg: grey;
+$fc: black;
+$bc: orange;
+$close: darkred;
+$border: white;
+$ld: mediumaquamarine;
