+++ public/colorsets/irie.scss
@@ -1,3 +1,6 @@
 $bg: black;
-$fc: darkgrey;
-$bc: white;
+$fc: yellow;
+$bc: green;
+$close: red;
+$border: green;
+$ld: yellow;
