+++ public/colorsets/modern.scss
@@ -1,3 +1,6 @@
 $bg: black;
-$fc: darkgrey;
-$bc: white;
+$fc: rgb(87, 87, 87);
+$bc: rgb(167, 167, 167);
+$close: hsl(0, 100%, 7%);
+$border: rgb(118, 118, 118);
+$ld: rgb(190, 255, 177);
