+++ public/layouts/mobil.scss
@@ -1,7 +1,20 @@
+input,
+select {
+  height: 44px;
+  border-radius: 9px;
+}
+
+button:hover {
+  background: white;
+  color: mediumaquamarine;
+}
+textarea {
+  height: 88px;
+  border-radius: 9px;
+}
 #nav {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   top: 0;
   left: 0;
@@ -24,12 +37,18 @@
   width: auto !important;
   text-align: left !important;
 }
+#dataBox,
+#upBox,
+#avtBox,
+#socioBox {
+  grid-column: 1/-1;
+  text-align: center;
+}
 
 #mobnav {
   display: none;
   width: 100%;
   height: calc(100% - 48px);
-  background: $bg;
   position: absolute;
   top: 24px;
   left: 0;
@@ -40,7 +59,6 @@
   grid-column: 1/-1;
   grid-row: 3;
   width: 100%;
-  color: $fc;
 }
 #closemob {
   grid-row: 2;
@@ -51,16 +69,10 @@
   height: 36px;
   font-size: 32px;
   border-radius: 50%;
-  background: $fc;
-  color: $bg;
   text-align: center;
   justify-self: center;
   overflow: hidden;
   transform: rotate(45deg);
-}
-#closemob:hover {
-  background: red;
-  color: $fc;
 }
 .mslink {
   width: 33%;
@@ -69,7 +81,6 @@
 }
 .thick1,
 .thick2 {
-  background: $bg;
   width: 12px;
   border-radius: 2px;
   height: 100%;
@@ -83,7 +94,6 @@
 
 #mobbtn {
   display: block;
-  background: $bg;
   width: 22px;
   height: 22px;
   grid-column: 5;
@@ -92,7 +102,6 @@
 #stage {
   width: 100%;
   height: calc(100% - 48px);
-  background: $fc;
   position: absolute;
   top: 24px;
   left: 0;
@@ -102,7 +111,6 @@
 #foot {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   bottom: 0px;
   left: 0;
@@ -140,16 +148,7 @@
 .mlink {
   font-size: 40px;
 }
-.mlink:hover,
-#contact:hover,
-#imprint:hover,
-#terms:hover,
-#mcontact:hover,
-#mimprint:hover,
-#mterms:hover {
-  background: $fc;
-  color: $bg;
-}
+
 #sublinks {
   margin-top: 2em;
 }
@@ -164,8 +163,6 @@
 }
 #admin,
 #madmin {
-  background: black !important;
-  color: white !important;
   width: 100%;
   height: 22px;
   position: absolute;
@@ -190,7 +187,6 @@ img {
 }
 .line {
   height: calc(100% - 4px);
-  background: darkgrey;
   width: 3px;
   border-radius: 1px;
   float: left;
