+++ public/layouts/pc.scss
@@ -1,7 +1,6 @@
 #nav {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   top: 0;
   left: 0;
@@ -23,8 +22,9 @@
   text-align: left !important;
 }
 
-#mobnav {
-  display: none;
+#mobnav,
+#mobNavList {
+  display: none !important;
 }
 #mobbtn {
   display: none;
@@ -38,34 +38,27 @@
 #contact {
   grid-column: 2;
 }
-#contact:hover,
-#imprint:hover,
-#terms:hover,
-#mcontact:hover,
-#mimprint:hover,
-#mterms:hover {
-  background: $fc;
-  color: $bg;
-}
-#admin,
-#madmin {
-  background: black !important;
-  color: white !important;
-}
+
 #stage {
   width: 100%;
   height: calc(100% - 48px);
-  background: $fc;
   position: absolute;
   top: 24px;
   left: 0;
   overflow-y: scroll;
   padding: 0 3em 0 3em;
 }
+#avtBox {
+  grid-row: 1/4;
+  text-align: center;
+}
+#submit_create {
+  grid-column: 3;
+  grid-row: 4;
+}
 #foot {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   bottom: 0px;
   left: 0;
