+++ public/layouts/tab.scss
@@ -1,7 +1,6 @@
 #nav {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   top: 0;
   left: 0;
@@ -25,19 +24,10 @@
 }
 
 #mobnav {
-  display: none;
+  display: none !important;
 }
 #mobbtn {
   display: none;
-}
-#contact:hover,
-#imprint:hover,
-#terms:hover,
-#mcontact:hover,
-#mimprint:hover,
-#mterms:hover {
-  background: $fc;
-  color: $bg;
 }
 #imprint {
   grid-column: 4;
@@ -52,8 +42,6 @@
 #madmin,
 #admin:hover,
 #madmin:hover {
-  background: black !important;
-  color: darkgrey !important;
   width: 40%;
   height: 22px;
   position: absolute;
@@ -63,17 +51,39 @@
 #stage {
   width: 100%;
   height: calc(100% - 48px);
-  background: $fc;
   position: absolute;
   top: 24px;
   left: 0;
   overflow-y: scroll;
   padding: 0 3em 0 3em;
 }
+#avtBox {
+  grid-row: 1/4;
+  grid-column: 3;
+  text-align: center;
+}
+#submit_create {
+  grid-column: 3;
+  grid-row: 4;
+}
+#socioBox,
+#dataBox,
+#upBox,
+#nftBox,
+#pfpBox,
+#picBox {
+  grid-column: 1/3;
+}
+#upBox,
+#nftBox,
+#pfpBox,
+#picBox {
+  grid-row: 3/-1;
+}
+
 #foot {
   width: 100%;
   height: 22px;
-  background: $fc;
   position: absolute;
   bottom: 0px;
   left: 0;
